replace atof with atoi
git-svn-id: http://svn.code.sf.net/p/rts-2/code/trunk/rts-2@5784 65ab8f4e-f147-0410-b3a9-f14133ecfe55

+++ src/utils/rts2valuearray.cpp
@@ -231,7 +231,7 @@ IntegerArray::setValueCharArr (const char *_value)
 	std::vector <std::string> sv = SplitStr (std::string (_value), std::string (" "));
 	for (std::vector <std::string>::iterator iter = sv.begin (); iter != sv.end (); iter++)
 	{
-		value.push_back (atof ((*iter).c_str ()));
+		value.push_back (atoi ((*iter).c_str ()));
 	}
 	changed ();
 	return 0;
